pyproject.toml
[project]
name = "d20"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "defusedxml",
    "fiftyone",
    "loguru>=0.7.0",
    "pillow",
    "pydantic>=2.0.0",
    "pyyaml",
    "requests>=2.31.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project.scripts]
d20 = "d20.cli:main"

[project.optional-dependencies]
dev = [
    "mypy",
    "pytest",
    "pytest-cov",
    "ruff",
]

[tool.hatch.build.targets.wheel]
packages = ["src/d20", "src/pillow"]

[tool.ruff]
target-version = "py310"
line-length = 100
select = [
    "E",    # pycodestyle errors
    "W",    # pycodestyle warnings
    "F",    # pyflakes
    "I",    # isort
    "B",    # flake8-bugbear
    "C4",   # flake8-comprehensions
    "UP",   # pyupgrade
    "ARG",  # flake8-unused-arguments
    "SIM",  # flake8-simplify
    "TCH",  # flake8-type-checking
    "PIE",  # flake8-pie
    "PT",   # flake8-pytest-style
    "RET",  # flake8-return
    "RUF",  # Ruff-specific rules
    "N",    # pep8-naming
    "S",    # flake8-bandit (security)
    "BLE",  # flake8-blind-except
    "FBT",  # flake8-boolean-trap
    "A",    # flake8-builtins
    "COM",  # flake8-commas
    "DTZ",  # flake8-datetimez
    "T20",  # flake8-print
    "FA",   # flake8-future-annotations
    "ISC",  # flake8-implicit-str-concat
    "ICN",  # flake8-import-conventions
    "PTH",  # flake8-use-pathlib
    "ERA",  # eradicate (commented-out code)
    "PD",   # pandas-vet
    "PGH",  # pygrep-hooks
    "PL",   # pylint
    "TRY",  # tryceratops
    "NPY",  # NumPy-specific rules
    "PERF", # perflint
    "FURB", # refurb
    "LOG",  # flake8-logging-format
    "RSE",  # flake8-raise
    "ASYNC", # flake8-async
    "TID",  # flake8-tidy-imports
    "Q",    # flake8-quotes
]
ignore = [
    "E501",   # line too long (handled by formatter)
    "S101",   # use of assert (allowed in tests)
    "PLR0913", # too many arguments
    "PLR2004", # magic value used in comparison
]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

[tool.ruff.lint.isort]
known-first-party = ["d20"]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]  # Allow unused imports in __init__.py

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true
show_error_codes = true
show_error_context = true
pretty = true
disallow_any_generics = true
disallow_subclassing_any = true
disallow_untyped_calls = true
disallow_untyped_expr = true
disallow_any_unimported = true
disallow_any_expr = false  # Too strict for practical use
disallow_any_decorated = false  # Too strict for practical use
warn_unused_coroutines = true
warn_incomplete_stub = true
warn_redundant_expr = true
strict_optional = true
strict = false  # Don't enable all strict flags, but enable most important ones

[[tool.mypy.overrides]]
module = [
    "tests.*",
]
disallow_untyped_defs = false

[tool.pytest.ini_options]
minversion = "8.0"
pythonpath = ["src"]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "-ra",
    "--strict-markers",
    "--strict-config",
    "--cov=d20",
    "--cov-report=term-missing",
    "--cov-report=html",
    "--cov-report=xml",
    "--cov-fail-under=80",
    "--maxfail=1",
    "--tb=short",
    "-W", "error::DeprecationWarning",
    "-W", "error::PendingDeprecationWarning",
]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
]